--- examples/0. get-start-with-your-level/5.streaming_pipelines/find_items_project.json
@@ -1,45 +1,0 @@
-{
-    "public_name": "Find a similar item in an online store 2",
-    "public_description": "Go to M&S online and find similar shoes on the website",
-    "task_spec": {
-        "input_spec": {
-            "image": {
-                "required": true,
-                "hidden": false,
-                "type": "url"
-            }
-        },
-        "output_spec": {
-            "found_link": {
-                "required": true,
-                "hidden": false,
-                "type": "string"
-            }
-        },
-        "view_spec": {
-            "localizationConfig": null,
-            "config": "{\n    \"view\": {\n        \"content\": {\n            \"url\": {\n                \"path\": \"image\",\n                \"type\": \"data.input\"\n            },\n            \"fullHeight\": true,\n            \"type\": \"view.image\"\n        },\n        \"controls\": {\n            \"items\": [\n                {\n                    \"content\": \"Find the same **shoes** on Marks and Spencer\",\n                    \"type\": \"view.markdown\"\n                },\n                {\n                    \"action\": {\n                        \"payload\": \"https://www.marksandspencer.com\",\n                        \"type\": \"action.open-link\"\n                    },\n                    \"label\": \"Marks and Spencer\",\n                    \"validation\": {\n                        \"url\": \"https://www.marksandspencer.com\",\n                        \"type\": \"condition.link-opened\"\n                    },\n                    \"type\": \"view.action-button\"\n                },\n                {\n                    \"content\": \"Shoes must be the same color and the same style.\",\n                    \"type\": \"view.text\"\n                },\n                {\n                    \"data\": {\n                        \"path\": \"found_link\",\n                        \"type\": \"data.output\"\n                    },\n                    \"label\": \"Paste the link here\",\n                    \"validation\": {\n                        \"conditions\": [\n                            {\n                                \"type\": \"condition.required\"\n                            },\n                            {\n                                \"data\": {\n                                    \"path\": \"found_link\",\n                                    \"type\": \"data.output\"\n                                },\n                                \"schema\": {\n                                    \"type\": \"string\",\n                                    \"pattern\": \"marksandspencer.com/?\"\n                                },\n                                \"hint\": \"the link must be from the Marks and Spencer website\",\n                                \"type\": \"condition.schema\"\n                            }\n                        ],\n                        \"type\": \"condition.all\"\n                    },\n                    \"type\": \"field.text\"\n                }\n            ],\n            \"type\": \"view.list\"\n        },\n        \"controlsWidth\": 320.0,\n        \"type\": \"layout.sidebar\"\n    },\n    \"plugins\": [\n        {\n            \"layout\": {\n                \"kind\": \"scroll\",\n                \"taskWidth\": 800.0\n            },\n            \"type\": \"plugin.toloka\"\n        }\n    ]\n}",
-            "type": "tb",
-            "lock": {
-                "core": "1.0.0",
-                "view.image": "1.0.0",
-                "view.markdown": "1.0.0",
-                "action.open-link": "1.0.0",
-                "condition.link-opened": "1.0.0",
-                "view.action-button": "1.0.0",
-                "view.text": "1.0.0",
-                "condition.required": "1.0.0",
-                "condition.schema": "1.0.0",
-                "condition.all": "1.0.0",
-                "field.text": "1.0.0",
-                "view.list": "1.0.0",
-                "layout.sidebar": "1.0.0",
-                "plugin.toloka": "1.0.0"
-            }
-        }
-    },
-    "assignments_issuing_type": "AUTOMATED",
-    "assignments_automerge_enabled": false,
-    "public_instructions": "\nLook at the shoes the person is wearing in the picture. \nGo to M&amp;S online store and find the same or similar pair of shoes on the website.\nThe shoes should be similar in color, style or height.\n",
-    "private_comment": "streaming_piplines example find_items_project"
-}
