--- examples/0. get-start-with-your-level/5.streaming_pipelines/verification_pool.json
@@ -1,93 +1,0 @@
-{
-    "private_name": "Is ok?",
-    "may_contain_adult_content": true,
-    "assignment_max_duration_seconds": 60,
-    "defaults": {
-        "default_overlap_for_new_task_suites": 1
-    },
-    "auto_close_after_complete_delay_seconds": 0,
-    "auto_accept_solutions": true,
-    "auto_accept_period_day": 14,
-    "assignments_issuing_config": {
-        "issue_task_suites_in_creation_order": false
-    },
-    "priority": 0,
-    "filter": {
-        "and": [
-            {
-                "operator": "IN",
-                "value": "EN",
-                "key": "languages",
-                "category": "profile"
-            }
-        ]
-    },
-    "quality_control": {
-        "configs": [
-            {
-                "rules": [
-                    {
-                        "action": {
-                            "parameters": {
-                                "scope": "ALL_PROJECTS",
-                                "duration": 3,
-                                "duration_unit": "DAYS",
-                                "private_comment": "Doesn't match the majority"
-                            },
-                            "type": "RESTRICTION_V2"
-                        },
-                        "conditions": [
-                            {
-                                "operator": "GTE",
-                                "value": 4,
-                                "key": "total_answers_count"
-                            },
-                            {
-                                "operator": "LT",
-                                "value": 75.0,
-                                "key": "correct_answers_rate"
-                            }
-                        ]
-                    }
-                ],
-                "collector_config": {
-                    "uuid": "8a34d3d8-fa33-4681-b728-194f5937a1c6",
-                    "parameters": {
-                        "answer_threshold": 3
-                    },
-                    "type": "MAJORITY_VOTE"
-                }
-            },
-            {
-                "rules": [
-                    {
-                        "action": {
-                            "parameters": {
-                                "delta": 1,
-                                "open_pool": true
-                            },
-                            "type": "CHANGE_OVERLAP"
-                        },
-                        "conditions": [
-                            {
-                                "operator": "EQ",
-                                "value": "REJECT",
-                                "key": "assessment_event"
-                            }
-                        ]
-                    }
-                ],
-                "collector_config": {
-                    "uuid": "1e5f1d39-a065-468a-95f3-36b7d0507be4",
-                    "type": "ASSIGNMENTS_ASSESSMENT"
-                }
-            }
-        ]
-    },
-    "mixer_config": {
-        "real_tasks_count": 1,
-        "golden_tasks_count": 0,
-        "training_tasks_count": 0
-    },
-    "type": "REGULAR"
-}
