--- examples/0. get-start-with-your-level/5.streaming_pipelines/verification_project.json
@@ -1,56 +1,0 @@
-{
-    "public_name": "Do the shoes look similar to each other?",
-    "public_description": "",
-    "task_spec": {
-        "input_spec": {
-            "image": {
-                "required": true,
-                "hidden": false,
-                "type": "url"
-            },
-            "found_link": {
-                "required": true,
-                "hidden": false,
-                "type": "url"
-            },
-            "assignment_id": {
-                "required": true,
-                "hidden": false,
-                "type": "string"
-            }
-        },
-        "output_spec": {
-            "result": {
-                "required": true,
-                "hidden": false,
-                "allowed_values": [
-                    "Yes",
-                    "No"
-                ],
-                "type": "string"
-            }
-        },
-        "view_spec": {
-            "localizationConfig": null,
-            "config": "{\n    \"view\": {\n        \"controls\": {\n            \"items\": [\n                {\n                    \"content\": \"Check that the uploaded image matches the product in the store.\",\n                    \"type\": \"view.text\"\n                },\n                {\n                    \"action\": {\n                        \"payload\": {\n                            \"path\": \"found_link\",\n                            \"type\": \"data.input\"\n                        },\n                        \"type\": \"action.open-link\"\n                    },\n                    \"label\": \"Check the item\",\n                    \"type\": \"view.action-button\"\n                },\n                {\n                    \"content\": \"Are these **shoes** similar to each other?\",\n                    \"type\": \"view.markdown\"\n                },\n                {\n                    \"content\": \"Shoes must be the same **color and the same style**\",\n                    \"type\": \"view.text\"\n                },\n                {\n                    \"data\": {\n                        \"path\": \"result\",\n                        \"type\": \"data.output\"\n                    },\n                    \"options\": [\n                        {\n                            \"value\": \"Yes\",\n                            \"label\": \"Yes\"\n                        },\n                        {\n                            \"value\": \"No\",\n                            \"label\": \"No\"\n                        }\n                    ],\n                    \"validation\": {\n                        \"type\": \"condition.required\"\n                    },\n                    \"type\": \"field.radio-group\"\n                }\n            ],\n            \"type\": \"view.list\"\n        },\n        \"items\": [\n            {\n                \"url\": {\n                    \"path\": \"image\",\n                    \"type\": \"data.input\"\n                },\n                \"fullHeight\": true,\n                \"type\": \"view.image\"\n            },\n            {\n                \"url\": {\n                    \"path\": \"found_link\",\n                    \"type\": \"data.input\"\n                },\n                \"fullHeight\": true,\n                \"type\": \"view.iframe\"\n            }\n        ],\n        \"type\": \"layout.side-by-side\"\n    }\n}",
-            "type": "tb",
-            "lock": {
-                "core": "1.0.0",
-                "view.text": "1.0.0",
-                "action.open-link": "1.0.0",
-                "view.action-button": "1.0.0",
-                "view.markdown": "1.0.0",
-                "condition.required": "1.0.0",
-                "field.radio-group": "1.0.0",
-                "view.list": "1.0.0",
-                "view.image": "1.0.0",
-                "view.iframe": "1.0.0",
-                "layout.side-by-side": "1.0.0"
-            }
-        }
-    },
-    "assignments_issuing_type": "AUTOMATED",
-    "assignments_automerge_enabled": false,
-    "public_instructions": "\nTake a look at 2 pictures. Decide whether the shoes look similar or not.\n<div><br />\n  <div>The shoes look similar if they are the same or <b>similar color, fabric and style</b>.</div>\n  <div>If you do not see any shoes in the pictures choose <b>No</b></div>\n</div>\n",
-    "private_comment": "streaming_piplines example verification_project"
-}
